county_health/requirements.txt
Django==1.10.3
django-crispy-forms==1.6.1
django-extensions==1.7.5
djangorestframework==3.5.3
Pillow==4.0.0
